+++ packages/sdk-ui-ts/src/utils/helpers.ts
@@ -5,5 +5,9 @@ export const getTokenLogoWithVendorPathPrefix = (image: string) => {
     return image
   }
 
+  if (image.includes('http')) {
+    return image
+  }
+
   return path.join('/', 'vendor', '@injectivelabs', 'token-metadata', image)
 }