+++ packages/tx-ts/CHANGELOG.md
@@ -3,6 +3,44 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.4](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/tx-ts@1.0.3...@injectivelabs/tx-ts@1.0.4) (2022-06-13)
+
+
+### Bug Fixes
+
+* broadcasting tx ([5816180](https://github.com/InjectiveLabs/injective-ts/commit/5816180ed2a5c5fdec61ef018d0ae7121f4f9072))
+
+
+
+
+
+## [1.0.3](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/tx-ts@1.0.2...@injectivelabs/tx-ts@1.0.3) (2022-06-13)
+
+
+### Features
+
+* txClient ([efab6c8](https://github.com/InjectiveLabs/injective-ts/commit/efab6c80190b44620cdce8aa20fed4fef213de79))
+
+
+
+
+
+## [1.0.2](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/tx-ts@1.0.1...@injectivelabs/tx-ts@1.0.2) (2022-06-13)
+
+**Note:** Version bump only for package @injectivelabs/tx-ts
+
+
+
+
+
+## [1.0.1](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/tx-ts@0.0.6...@injectivelabs/tx-ts@1.0.1) (2022-06-13)
+
+**Note:** Version bump only for package @injectivelabs/tx-ts
+
+
+
+
+
 ## [0.0.6](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/tx-ts@0.0.5...@injectivelabs/tx-ts@0.0.6) (2022-06-12)
 
 **Note:** Version bump only for package @injectivelabs/tx-ts