+++ packages/wallet-ts/CHANGELOG.md
@@ -3,6 +3,66 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.6](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/wallet-ts@1.0.5...@injectivelabs/wallet-ts@1.0.6) (2022-06-13)
+
+
+### Bug Fixes
+
+* broadcasting tx ([5816180](https://github.com/InjectiveLabs/injective-ts/commit/5816180ed2a5c5fdec61ef018d0ae7121f4f9072))
+
+
+
+
+
+## [1.0.5](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/wallet-ts@1.0.4...@injectivelabs/wallet-ts@1.0.5) (2022-06-13)
+
+
+### Features
+
+* txClient ([efab6c8](https://github.com/InjectiveLabs/injective-ts/commit/efab6c80190b44620cdce8aa20fed4fef213de79))
+
+
+
+
+
+## [1.0.4](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/wallet-ts@1.0.3...@injectivelabs/wallet-ts@1.0.4) (2022-06-13)
+
+
+### Features
+
+* import web3 in the sdk-ui-ts ([12bbb93](https://github.com/InjectiveLabs/injective-ts/commit/12bbb936c5e057abdaa89127f5c65fcd9f5f1572))
+
+
+
+
+
+## [1.0.3](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/wallet-ts@1.0.2...@injectivelabs/wallet-ts@1.0.3) (2022-06-13)
+
+**Note:** Version bump only for package @injectivelabs/wallet-ts
+
+
+
+
+
+## [1.0.2](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/wallet-ts@1.0.1...@injectivelabs/wallet-ts@1.0.2) (2022-06-13)
+
+
+### Bug Fixes
+
+* alchemy version ([b3cdb5f](https://github.com/InjectiveLabs/injective-ts/commit/b3cdb5fd3aaf0feb51332af65749d113ac666c5c))
+
+
+
+
+
+## [1.0.1](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/wallet-ts@0.0.31...@injectivelabs/wallet-ts@1.0.1) (2022-06-13)
+
+**Note:** Version bump only for package @injectivelabs/wallet-ts
+
+
+
+
+
 ## [0.0.31](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/wallet-ts@0.0.30...@injectivelabs/wallet-ts@0.0.31) (2022-06-12)
 
 **Note:** Version bump only for package @injectivelabs/wallet-ts